Fix pandas 2.1.0 FutureWarnings (#226)
Fixes #224 and a bunch of other pandas 2.1.0 FutureWarnings.

+++ petab/calculate.py
@@ -83,7 +83,7 @@ def calculate_residuals_for_table(
     # create residual df as copy of measurement df, change column
     residual_df = measurement_df.copy(deep=True).rename(
         columns={MEASUREMENT: RESIDUAL})
-
+    residual_df[RESIDUAL] = residual_df[RESIDUAL].astype("float64")
     # matching columns
     compared_cols = set(MEASUREMENT_DF_COLS)
     compared_cols -= {MEASUREMENT}
